politics_backend/requirements.txt
django>=5.0
djangorestframework>=3.15.0
django-cors-headers>=4.0.0
python-decouple>=3.8
appwrite>=14.1.0
